IPython console: Move ipyconsole fixture to conftest
